refactor: template manipulation functions
